requirements.txt
Django==5.0.*
mysqlclient==2.2.*
gunicorn==21.*
celery==5.3.*
redis==5.0.*
django-environ==0.11.*
python-dotenv==1.0.*
stripe==11.1.0
Pillow>=10.0.0
